backend/requirements.txt
fastapi
uvicorn[standard]
requests
python-dotenv
pytest
gunicorn
pyodbc